backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.10
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
CacheControl==0.14.4
cachetools==6.2.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cryptography==46.0.3
deprecation==2.1.0
fastapi==0.115.0
firebase_admin==7.1.0
frozenlist==1.8.0
google-api-core==2.28.1
google-auth==2.43.0
google-cloud-core==2.5.0
google-cloud-firestore==2.21.0
google-cloud-storage==3.6.0
google-crc32c==1.7.1
google-resumable-media==2.8.0
googleapis-common-protos==1.72.0
gotrue==2.12.4
grpcio==1.76.0
grpcio-status==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
http_ece==1.2.1
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
msgpack==1.1.2
multidict==6.7.0
packaging==25.0
postgrest==0.18.0
propcache==0.4.1
proto-plus==1.26.1
protobuf==6.33.1
py-vapid==1.9.2
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.9.2
pydantic_core==2.23.4
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pywebpush==2.1.2
PyYAML==6.0.3
realtime==2.5.3
requests==2.32.5
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
starlette==0.38.6
storage3==0.9.0
supabase==2.10.0
supafunc==0.7.0
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.32.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==15.0.1
yarl==1.22.0
